Let random maximum population changes not affect Major home systems.
I consider this a bug and overseight. It can ruin one's game in case
there's no other system with large enough population to build
the type 3 shipyard, while there are too many colonies making it
unlikely the homesystem gets its maximum population increased, while
it is rather likely to be decreased in the early game due to missing
colonies.