implement a warning message if a fleet will waste terraform points
This has bothered me since long. Now it's no longer needed to check
terraforming fleets whether they would waste points due to the terraforming
being finished. Works only right if all ships terraforming a certain
planet are all in the same fleet. So if other single ships or fleets
terraform the same planet it can be wrong, due to how the ship
commands are processed in general. (It would require much more
complicated code.) There may also be ships of other players terraforming
the same planet and we don't know what that player is going to do, so
a prediction can never be 100% correct.