CBotf2Doc::m_Sector and ::m_System memory structure refactoring step #5
Rename the variables. Manually fix remaining compile errors.

This completes the change for both variables from
std::vector<std::vector> to a single std::vector, which is accessed
by y*starmap_width + x, that means, number of times we already walked
through the width of the starmap plus the current x position, starting
with (0,0) in the top left corner.

reasons for doing this:
-I had the impression that the way of initializing and keeping the
systems and sectors may be the cause of some memmory problems - it seems
this suspicion unfortunately was wrong.
-It is however anyway better to have only a single vector which means
that the according memory is allocated in a single block as well instead
of scattered in pieces, which slows down the access. This should outweight
the disadvantage that each (most) of the time it is accessed there is now
the additional calculation y*starmap_width + x, whose occurances
can be reduced over time when the according places get refactored.

To be able to get this done in time at all I wrote a small script which
did most of the zillions of modifications, with the drawback of the
addtional calculations.