started to eliminate cases where CShip::m_TargetKO[0] is set to CShip::m_KO
In order to unset a ship's target, since the fact that a ship can have no
target while CShip::m_TargetKO[0] is set to CPoint(-1, -1) OR
CShip::m_KO was the root cause of the crashes when grouping ships.
So, let's ensure that ships with no target always have
CShip::m_TargetKO[0]=CPoint(-1, -1).

This commit affects the ship ai - not particularly sure about these.