drastically reduce xp gainable via ship-explore random event
It is now roughly the same as can be get by a few rounds of ship
training (with most major's or some minor's ship training building),
and staying 10 rounds there yields at least the same result.
(10 rounds can be lot sometimes...)
It seems to me that this is the cause of the lots of recent "strange
fights" which should have been won (or lost) with 90+% but weren't
and is also probably very much an advantage for races that start
with fighting ships.
The random events in general seem unbalanced.