new shipname generation algorithm
This should fix the duplicate Morlock raider warnings. Also was pretty
bad since at many places ship names are used to uniquely identify a ship.

Ship names were not really unique which was noticable on this minor with
many ships since it doesn't have ship names and only gets letters,
it started from A again after Z was reached. Now it continues with
27 upwards (1-26 match A-Z).

SAVEGAMES ARE INCOMPATIBLE.