ship controls improvement
When setting a ship target, active orders are no longer automatically
unset now. This means that the cloak command is no longer unset
when setting a target; same for other commands which allow to be
combined with the ship having an active moveto target. Thus the order
of giving the commands is unimportant.