fix ships in a fleet having ships in their fleets/ship duplication
Still not sure about best design there.
ake the parameter const again.
Duplice ship test should take care for that the caller deletes
the ships from the fleet of the passed ship afterwards (so that we do
not need to do it within this function). Actually, all callers appear
to delete the passed ship afterwards which of course makes sense.
Caught by sanity checks. :)