reduce potentially unsafe caching of CShips objects
expressions of the form:
CShips* pShip = &i->second;
and
CShips& ship = i->second;

If the iterator this came from invalidates, this pointer/reference
does as well. The iterator itself my be moved previously however and stay
valid.
Problem affects also passing i->second to a function, if it takes a
reference or pointer to the CShips.
Need to be careful when deleting ships (while adding is save since
the change to a map).