refactor relation between m_nTerraformingPlanet and m_iCurrentOrder in CShip(s)
There are crashes triggered by the terraform overlay, due to
inconsistencies when m_nTerraformingPlanet == -1 but still
m_iCurrentOrder == TERRAFORM.
Now it should be enforced that m_nTerraformingPlanet == -1
if and only if
m_iCurrentOrder == TERRAFORM

Apparently it is or was also supported at some spot that a single
planet can be colonized (as opposed to all colonizable planets in a
system) but shipbottomview doesn't contain code to support it.
Avoid this unnecessary complexity.