ensure consistent orders in a Chips
If any added ship cannot adopt the leader's order, forcefully unset
the leader's order instead of leaving the fleet with inconsistent orders.
This matches the fact that the order also can't be given *after*
the non-capable ship has been added (for most orders), and removes
exceptions I added in recent revisions because adding transporters
to terraforming fleets and such is allowed, but the transporter
cannot take over the terraform order.