logically unlink SHIP_ORDER::ATTACK/AVOID from combat tactics
Follows the logic of the previous commit.
These two no longer appear to the user; they are only used internally
to signal that COMBAT_TACTIC::ATTACK/AVOID is given.
This removes the semantic ambiguousness attack/avoid had; setting
the ship to attack/avoid on one hand, but also signalling that
it is "idle".