reduce amount of calls to CShips::HasFleet()
These calls were neccessary back when the m_Fleet member was a pointer
and alloc'd on the heap, but now in all these cases the condition
is sufficiently covered by the control condition of the loop blocks
they were guarding.
This commit only removes HasFleet() calls; it unfortunately doesn't look
clean due to reducing indentation of the guarded blocks (but actually
should).