simplify assigning and unassigning of flagship status
The command is now in the "immediate orders" category and behaves like it,
no turn change needed any more. The flagship status of a ship can now
be unassigned by clicking onto "cancel". This is a new feature, it was not
possible to again have no flagship at all after assigning the status
to any ship previously.
Less hazzle for the player, AI was and is not using flagships, the
overcrowded single turn orders category is less populated in favor
of the empty immediate category, and the related code simplifies a lot.