Allow to combine building prerequisites
http://www.forum.birth-of-the-empires.de./viewtopic.php?p=32478#p32478

It is only possible to enable one building prerequisite per building.
* only homesystem
* only own colony
* only minorrace stystem
* only conquered system
As a result the activation of multiple prerequisites render a Building
not buildable in any system.

Change that behavior to allow prerequisite combinations.

System.cpp:
Modify to check all combinations if required, and mark a building as
buildable if one of the prerequisites is fulfilled.
Therefore make every combination of building prerequisites possible.

StringTable.txt and SystemMenuView.cpp:
Add "ONLY_MINORRACE_SYSTEM" string for implementation of GUI text for
the "minor race system" prerequisite.

Buildings.data:
Change the affected buildings to fit the new prerequisite rules.
No change on the functioning of the existing buildings.