Made it so that previous stations are regularly disassembled on upgrade.
This destroys upgraded outposts/starbases (their previous versions, once
that the upgrade is finished) using the same logic (and code) as if
they had been regularly disassembled, so that boni for resources
and gold are gained in case the station was over an owned system.
Otherwise upgrading stations via the upgrade command suffers from a
severe disadvantage over the traditional method via disassembling
and rebuilding.