simplify taking energy buildings online/offline if manager active
The question whether we can take online a building in systems energy
menu (and it gets green) has changed from "do we produce enough energy ?"
to "could we produce enough energy (by eventually redistributing workers) ?"
That way, the manager can immidiately react with workers to what we set
on/offline. It's no longer needed to annoyingly provide energy before
taking something online, allowing the manager to take away the energy
if we take something offline.
If auto-taking buildings on/offline is activated, clicking one of the
handled types doesn't seem to have effect, as the manager takes it
immediately on/offline afterwards.