make CSystem a subclass to CSector
A system is always also a sector, isn't it.
This already simplified arguments to lots of functions.
A lot of functions in CSystem take things as parameters that are
available in CSector. These functions force their callers to make the
parameters available just for passing them up. CSystem/CSector have
the redundant owner field, and I often wished to have a coordinates member
in CSystem. We already allocate a CSystem for every map tile even though
it should be neccessary for the sun systems only. And finally, I
got rid of CBotEDoc::m_Sectors.
So, there's a vector of systems only now, each of which also has its
sector mother object constructed.