change type of owner member in CMapTile to CRace
This simplifies code, since very often, after getting the owner ID,
a call to CRaceController to get the according race follows.
Using boost:shared_ptr prevents the possibility of a dangling CRace
pointer owner member variable in CMapTile. (It would still be a bug,
but we can dereference and check for it.)