made it so that the self-compiled network lib form BotE Tools is linked in
Remove redundant versions of foundation, network and LZMA headers.
Use the headers from network lib.

If you get linker errors, open BotE Tools\NetworkLib\BotE Network.sln
and build that. It needs to be done for debug and release separately.
This change should allow debugging into the network lib independently
from who it does, as the paths to the cpp files are relative.