add missing boost headers
The folder in which these are is named "debug" which is ignored,
since the build output folder is named "Debug" and that one needs to be
ignored (windows case insensitivity). Thus these files
didn't show up as added/changed and I forgot to add them.

I hope it's now buildable on systems other than mine. (again)