feature: implement "forcing of minors"
This adds a field to BotE.ini in the [SPECIALS] section of the form
MINORS=<name>
or
NIMORS=<name>,<name>
No spaces, it needs to be the home system name.
example:
MINORS=Ehlen,Delari,Vulmar

These minors will be the first ones which are generated. It can happen
that not all of them are generated if minor density and/or system count
is too low.