worsen relation when bombarding a minor (to that minor)
I remember this as a bug being mentioned, and indeed it seems
it's still there. So, you could bomb a minor almost to death but still
be on max relation.
I wonder how this turns out; it's possible that AI plays worse, but
also that it plays better. This toggling between presents and bombing
of the same minor is always crazy. If it just forgets about presents
if bombing at all since relation is so bad then - good.