refactor the way sector names are displayed
Almost entirely removed letters, they only remain for anomalies where
it's kind of a nice feature. Coords are displayed in the form (x|y).
Coords are now 0-based, same as in the data saved in the code, so most
top-left is (0|0) and most bottom-right is (width-1|height-1).
Ship target data is displayed more conveniently in the form <name> (x|y).
Target data of friendly ships is displayed as well (needed: affiliation
with the owner).