implement defining exceptions for manager taking buildings on/offline
The manager will never take the defined exceptions on- or offline
but leave them for manual handling instead.
There are buildings for which it is way too hard for the manager
to decide whether they should be online, which forces the player
to not build them or to not use automatic on/offlining in that system
(until now).