when upgrading buildings, keep their on/offline status
Was particularly annoying in the case of defense buildings and becomes
annoying again in the case of energy buildings ignored by manager,
such as continuum scanner in an area where there are enough
alternative sources of scanning power.
If there are several of the same type, the last instance determines
the status, but the on/off rules don't differ.