made it so that the self-compiled lzma lib is used
The build process is now:
1. build lzma lib project
2. build network lib project
3. build BotE itself

In the long run, we should merge the three into a single sln (wile keeping
them in single projects).

This has the following effects:
-we can computerindependently debug into lzma lib
-cppcheck and memory checker tools can improve lzma lib
-the linker warnings about missing debug info are gone - we are at 0 :)
-compilation with VS2010 (and possibly newer) is supposed to work.
To compile with VS2010 (or newer), one has to use the upgrade assistant
for the three projects mentioned above and compile them in said order.
No guarantee that there won't be any further problems; however, when
I last tried vc10 the lzma lib being compiled with vc8 seemed to be
the only problem.