add modeless dialog box without static control
a modeless dialog box for tutorial is implemented which currently only
shows a default text line. a static or edit control for text display
needs to be added; a code snippet (commented out because it doesn't
work) using a new afx thread for bypassing modal afxmessagebox call is
added to StdAfx.h. Maybe someone can make this workaround work.