refactor CStatistics class
savegames incompatible

Most stuff is calculated at turn change now. Iterations over systems
should be much fewer.
Ranks are now calculated so that the rank is higher (worse) only if
the value is actually worse than the better major and not also if
it is equal. This was particularly a problem in the beginning of the game
with moral stats, for majors that come "last" (Omega for instance)