make marks calculation more precise
The best mark (1) is given to the best major, the worst mark
(count_of_majors) to the worst, everything else is mapped according
to those. So, if, for instance, a certain major is way ahead in a certain
category, there is now no mayor with mark 2,something but the second best
has mark 4 only (for instance). This is fairer than previously, when the
second best would just get mark (rank) 2 anyway.